requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==3.0.1
distlib==0.3.6
Django==4.1.3
django-mathfilters==1.0.0
filelock==3.8.0
idna==3.4
Pillow==9.4.0
platformdirs==2.5.2
python-dotenv==0.21.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.14
virtualenv==20.16.5
virtualenv-clone==0.5.7
